Add files via upload
Last week of Feb and as of right now March 2, I dedicated to coding and designing the project prototype(MVP) both the front and back end. On the front end, i am running an html styled by styles.css, hosted by your machine's local server, functionality supported by script.js. Script.js communicates with the backend which is comprised of two parts, server.py and llama.py. the backend is ran on a local flask server which communicates with two apis, hugging face and llama 2 chat. This is missing some functionalities however it should give yall a start. Read the README to see commands to run this project.